Yushi's Code Review
Great project! I like the getSearch function that can retrieve books from the database based on a search query.
The codes are well-organized and placed in the appropriate folders, but it might be better to remove the unused code in the search.js files. And also may need to include the LICENSE file in the repository.